module refactoring part 2
